chore: display snapping point only when mouse is in the renge of the snap tolerance
