+++ index.html
@@ -15,10 +15,10 @@
     <meta charset="utf-8" />
   </head>
   <body>
-    <script src="https://cdn.jsdelivr.net/npm/ol@7.1.0/dist/ol.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/ol@7.3.0/dist/ol.js"></script>
     <script src="https://unpkg.com/maplibre-gl@2.4.0/dist/maplibre-gl.js"></script>
     <script src="https://unpkg.com/jsts@2.6.1/dist/jsts.min.js"></script>
-    <script src="build/index.js"></script>
+    <script src="index.js"></script>
     <div id="app">
       <div id="header">
         <nav>
