+++ src/control/index.js
@@ -1,21 +1,10 @@
-import Control from './control';
-import CAD from './cad';
-import Rotate from './rotate';
-import Draw from './draw';
-import Modify from './modify';
-import Buffer from './buffer';
-import Union from './union';
-import Intersection from './intersection';
-import Difference from './difference';
-
-export {
-  Control,
-  CAD,
-  Rotate,
-  Draw,
-  Modify,
-  Buffer,
-  Union,
-  Intersection,
-  Difference,
-};
+export { default as Control } from './control';
+export { default as CAD } from './cad';
+export { default as Rotate } from './rotate';
+export { default as Draw } from './draw';
+export { default as Modify } from './modify';
+export { default as Buffer } from './buffer';
+export { default as Union } from './union';
+export { default as Intersection } from './intersection';
+export { default as Difference } from './difference';
+export { default as Toolbar } from './toolbar';
