+++ src/event/index.js
@@ -1,0 +1,6 @@
+export { default as DeleteEvent } from './delete-event';
+export { default as MoveEvent } from './move-event';
+export { default as SnapEvent } from './snap-event';
+export * from './snap-event';
+export * from './move-event';
+export * from './delete-event';
