+++ src/interaction/index.js
@@ -1,6 +1,4 @@
-import Delete from './delete';
-import SelectMove from './selectmove';
-import SelectModify from './selectmodify';
-import Move from './move';
-
-export default { Delete, SelectModify, SelectMove, Move };
+export { default as Delete } from './delete';
+export { default as SelectMove } from './selectmove';
+export { default as SelectModify } from './selectmodify';
+export { default as Move } from './move';
