+++ src/service/index.js
@@ -1,4 +1,2 @@
-import LocalStorage from './local-storage';
-import Storage from './storage';
-
-export { Storage, LocalStorage };
+export { default as LocalStorage } from './local-storage';
+export { default as Storage } from './storage';
